// src/Agent/Agent.Portal/Client/src/Common/Constants/ApiVersions.ts
/* eslint-disable @typescript-eslint/no-duplicate-enum-values */

export enum ApiVersions {
    armBatchApiVersion20250301 = '2025-03-01',
    microsoftAppApiVersion20250501Preview = '2025-05-01-preview',
    appServiceApiVersion20250301 = '2025-03-01',
    tenantsApiVersion20221201 = '2022-12-01',
    subscriptionsApiVersion20200101 = '2020-01-01',
    armApiVersion20230301 = '2023-03-01',
    permissionsApiVersion20220401 = '2022-04-01',
    providerApiVersion20231101 = '2023-11-01',
    listLocationApiVersion20200101 = '2020-01-01',
    resourceApiVersion20200101 = '2020-01-01',
    resourceProviderApiVersion20220901 = '2022-09-01',
    resourceGroupApiVersion20200601 = '2020-06-01',
    argQueryApiVersion20240401 = '2024-04-01',
    appInsightsApiVersion20220615 = '2022-06-15',
    identityApiVersion20241130 = '2024-11-30',
    workspacesApiVersion20200801 = '2020-08-01',
    userIdentityApiVersion20181130 = '2018-11-30',
}