// src/Agent/Agent.Portal/Client/src/Common/Contracts/Permissions.ts
export type Permission = {
    actions: string[];
    notActions: string[];
    dataActions: string[];
    notDataActions: string[];
};

export type RoleDefinition = {
    roleName: string;
    type: string;
    permissions: Permission[];
    description?: string;
    assignableScopes?: string[];
};

export type RoleAssignment = {
    createdOn: string;
    roleDefinitionId: string;
    updatedBy: string;
    scope: string;
    principalId?: string;
    principalType?: string;
};

export enum LockLevels {
    readOnly = 'ReadOnly',
    canNotDelete = 'CanNotDelete',
}

export type Lock = {
    level: string;
    notes: string;
};

export type DenyAssignment = {
    principalId: string;
    permissions: {
        actions: string[];
        notActions: string[];
        dataActions: string[];
        notDataActions: string[];
    };
    scope: string;
    isSystemProtected: boolean;
};
// ...
export type PolicyEvaluationResult = {
    effectDetails: PolicyEffectDetails;
    evaluationDetails: CheckRestrictionEvaluationDetails;
    evaluationResult: string;
    policyInfo: PolicyReference;
};

export type ContentEvaluationResult = {
    policyEvaluations: PolicyEvaluationResult[];
};

export type PolicyReference = {
    policyAssignmentId: string;
    policyDefinitionId: string;
    policyDefinitionReferenceId: string;
    policySetDefinitionId: string;
};

export enum FieldRestrictionResult {
    Required = 'Required',
    Deny = 'Deny',
    Audit = 'Audit',
    Removed = 'Removed',
}

export type FieldRestriction = {
    defaultValue: string;
    policy: PolicyReference;
    policyEffect: string;
    reason: string;
    result: FieldRestrictionResult;
    values: string[];
};

export type FieldRestrictions = {
    field: string;
    restrictions: FieldRestriction[];
};

export type PolicyAssignment = {
    contentEvaluationResult: ContentEvaluationResult;
    fieldRestrictions: FieldRestrictions[];
};

export enum PermissionPrincipalType {
    servicePrincipal = 'ServicePrincipal',
    user = 'User',
}

export enum RoleDefinitionId {
    Contributor = `/providers/Microsoft.Authorization/roleDefinitions/b24988ac-6180-42a0-ab88-20f7382dd24c`,
    AcrPull = `/providers/Microsoft.Authorization/roleDefinitions/7f951dda-4ed3-4680-a7ca-43fe172d538d`,
}
export interface PermissionsCheckResponse {
    value: [
        {
            actions: string[];
            notActions: string[];
            dataActions: string[];
            notDataActions: string[];
        },
    ];
}

export enum PermissionAction {
    read = './read',
    write = './write',
    stopSiteAction = './stop/action',
    startSiteAction = './start/action',
}

export interface Permissions {
    /**
     * The actions that are permitted.
     */
    readonly actions: ReadonlyArray<string>;

    /**
     * The actions that are explicitly not permitted.
     */
    readonly notActions: ReadonlyArray<string>;
}